src/pages/api/posts/AnalyticsController.test.ts
import { describe, it, expect } from "vitest";
import trackEvent from "./AnalyticsController";

describe("AnalyticsController", () => {
  it("should track an analytics event", async () => {
    const analyticsData = {
      pagevisited: "home",
      trafficsource: "search engine",
      timespentonpages: { home: 120 },
      userip: "192.168.0.1",
      browserinfo: "Chrome 94.0",
      useraction: "clicked button",
      campaignreference: "campaign1",
      conversionbehavior: { "newsletter signup": true },
    };
    const result = await trackEvent(analyticsData);
    expect(result).toBeTruthy();
  });

  it("should handle empty analytics data", async () => {
    const analyticsData = {};
    const result = await trackEvent(analyticsData);
    expect(result).toBeFalsy();
  });

  it("should handle invalid analytics data format", async () => {
    const analyticsData = {
      pagevisited: 123, // invalid data type
      trafficsource: "search engine",
      timespentonpages: { home: 120 },
      userip: "192.168.0.1",
      browserinfo: "Chrome 94.0",
      useraction: "clicked button",
      campaignreference: "campaign1",
      conversionbehavior: { "newsletter signup": true },
    };
    const result = await trackEvent(analyticsData);
    expect(result).toBeFalsy();
  });

  it("should handle errors during tracking", async () => {
    const analyticsData = {
      pagevisited: "home",
      trafficsource: "search engine",
      timespentonpages: { home: 120 },
      userip: "192.168.0.1",
      browserinfo: "Chrome 94.0",
      useraction: "clicked button",
      campaignreference: "campaign1",
      conversionbehavior: { "newsletter signup": true },
    };
    // assuming an error scenario
    const result = await trackEvent(analyticsData);
    expect(result).toBeFalsy();
  });
});
